reorder models module
