+++ ms2deepscore/models/__init__.py
@@ -1,0 +1,7 @@
+from .load_model import load_model
+from .SiameseModel import SiameseModel
+
+__all__ = [
+    "load_model",
+    "SiameseModel",
+]
